move get_data to Data object
