+++ specparam/data/data.py
@@ -7,6 +7,7 @@ import numpy as np
 
 from specparam.sim.gen import gen_freqs
 from specparam.data import SpectrumMetaData, ModelChecks
+from specparam.utils.array import unlog
 from specparam.utils.spectral import trim_spectrum
 from specparam.utils.checks import check_input_options
 from specparam.reports.strings import gen_data_str
@@ -155,6 +156,54 @@ class Data():
         """
 
         return SpectrumMetaData(**{key : getattr(self, key) for key in self._meta_fields})
+
+
+    def get_data(self, component='full', space='log'):
+        """Get a data component.
+
+        Parameters
+        ----------
+        component : {'full', 'aperiodic', 'peak'}
+            Which data component to return.
+                'full' - full power spectrum
+                'aperiodic' - isolated aperiodic data component
+                'peak' - isolated peak data component
+        space : {'log', 'linear'}
+            Which space to return the data component in.
+                'log' - returns in log10 space.
+                'linear' - returns in linear space.
+
+        Returns
+        -------
+        output : 1d array
+            Specified data component, in specified spacing.
+
+        Notes
+        -----
+        The 'space' parameter doesn't just define the spacing of the data component
+        values, but rather defines the space of the additive data definition such that
+        `power_spectrum = aperiodic_component + peak_component`.
+        With space set as 'log', this combination holds in log space.
+        With space set as 'linear', this combination holds in linear space.
+        """
+
+        if not self.has_data:
+            raise NoDataError("No data available to fit, can not proceed.")
+        assert space in ['linear', 'log'], "Input for 'space' invalid."
+
+        if component == 'full':
+            output = self.power_spectrum if space == 'log' \
+                else unlog(self.power_spectrum)
+        elif component == 'aperiodic':
+            output = self._model.results.model._spectrum_peak_rm if space == 'log' else \
+                unlog(self.power_spectrum) / unlog(self._model.results.model._peak_fit)
+        elif component == 'peak':
+            output = self._model.results.model._spectrum_flat if space == 'log' else \
+                unlog(self.power_spectrum) - unlog(self._model.results.model._ap_fit)
+        else:
+            raise ValueError('Input for component invalid.')
+
+        return output
 
 
     def plot(self, plt_log=False, **plt_kwargs):
