+++ specparam/models/base.py
@@ -2,7 +2,6 @@
 
 from copy import deepcopy
 
-from specparam.utils.array import unlog
 from specparam.utils.checks import check_array_dim
 from specparam.modes.modes import Modes
 from specparam.modutils.errors import NoDataError
@@ -64,54 +63,6 @@ class BaseModel():
 
         if getattr(self, 'algorithm', None):
             self.algorithm._reset_subobjects(modes=self.modes, results=self.results)
-
-
-    def get_data(self, component='full', space='log'):
-        """Get a data component.
-
-        Parameters
-        ----------
-        component : {'full', 'aperiodic', 'peak'}
-            Which data component to return.
-                'full' - full power spectrum
-                'aperiodic' - isolated aperiodic data component
-                'peak' - isolated peak data component
-        space : {'log', 'linear'}
-            Which space to return the data component in.
-                'log' - returns in log10 space.
-                'linear' - returns in linear space.
-
-        Returns
-        -------
-        output : 1d array
-            Specified data component, in specified spacing.
-
-        Notes
-        -----
-        The 'space' parameter doesn't just define the spacing of the data component
-        values, but rather defines the space of the additive data definition such that
-        `power_spectrum = aperiodic_component + peak_component`.
-        With space set as 'log', this combination holds in log space.
-        With space set as 'linear', this combination holds in linear space.
-        """
-
-        if not self.data.has_data:
-            raise NoDataError("No data available to fit, can not proceed.")
-        assert space in ['linear', 'log'], "Input for 'space' invalid."
-
-        if component == 'full':
-            output = self.data.power_spectrum if space == 'log' \
-                else unlog(self.data.power_spectrum)
-        elif component == 'aperiodic':
-            output = self.results.model._spectrum_peak_rm if space == 'log' else \
-                unlog(self.data.power_spectrum) / unlog(self.results.model._peak_fit)
-        elif component == 'peak':
-            output = self.results.model._spectrum_flat if space == 'log' else \
-                unlog(self.data.power_spectrum) - unlog(self.results.model._ap_fit)
-        else:
-            raise ValueError('Input for component invalid.')
-
-        return output
 
 
     def print_settings(self, description=False, concise=False):
